Implement Camera and Frame classes
The Camera class implements the functionality of grabbing a Frame
from a device (or recording) and applying fixed corrections to it.
Internally, the object uses OpenCV for the capture and processing of
the images.

For now the only implemented correction is a rotation by a fixed angle.
Camera calibration corrections will be implemented at a later stage.

The Frame class is a thin wrapper around OpenCV's Mat, used to hide
the internal data structure when passing Frames between parts of the
application.